previous_code/tf_version/requirement.txt
tensorflow-gpu
opencv-python
matplotlib
pandas
numpy
sklearn
scipy
